+++ azure-servicebus/azure/servicebus/__init__.py
@@ -20,6 +20,8 @@ from .constants import (
     AZURE_SERVICEBUS_NAMESPACE,
     AZURE_SERVICEBUS_ACCESS_KEY,
     AZURE_SERVICEBUS_ISSUER,
+    SERVICE_BUS_HOST_BASE,
+    DEFAULT_HTTP_TIMEOUT,
 )
 
 from .models import (
