+++ azure-servicebus/azure/servicebus/_common_error.py
@@ -1,0 +1,88 @@
+﻿#-------------------------------------------------------------------------
+# Copyright (c) Microsoft.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#--------------------------------------------------------------------------
+from azure.common import (
+    WindowsAzureError,
+    WindowsAzureConflictError,
+    WindowsAzureMissingResourceError,
+)
+
+
+_ERROR_CONFLICT = 'Conflict ({0})'
+_ERROR_NOT_FOUND = 'Not found ({0})'
+_ERROR_UNKNOWN = 'Unknown error ({0})'
+_ERROR_VALUE_NONE = '{0} should not be None.'
+_ERROR_ASYNC_OP_FAILURE = 'Asynchronous operation did not succeed.'
+_ERROR_ASYNC_OP_TIMEOUT = 'Timed out waiting for async operation to complete.'
+_ERROR_MESSAGE_NOT_PEEK_LOCKED_ON_DELETE = \
+    'Message is not peek locked and cannot be deleted.'
+_ERROR_MESSAGE_NOT_PEEK_LOCKED_ON_UNLOCK = \
+    'Message is not peek locked and cannot be unlocked.'
+_ERROR_EVENT_HUB_NOT_FOUND = 'Event hub was not found'
+_ERROR_QUEUE_NOT_FOUND = 'Queue was not found'
+_ERROR_TOPIC_NOT_FOUND = 'Topic was not found'
+_ERROR_SERVICEBUS_MISSING_INFO = \
+    'You need to provide servicebus namespace, access key and Issuer'
+_WARNING_VALUE_SHOULD_BE_BYTES = \
+    'Warning: {0} must be bytes data type. It will be converted ' + \
+    'automatically, with utf-8 text encoding.'
+_ERROR_VALUE_SHOULD_BE_BYTES = '{0} should be of type bytes.'
+_ERROR_VALUE_NEGATIVE = '{0} should not be negative.'
+
+
+def _general_error_handler(http_error):
+    ''' Simple error handler for azure.'''
+    if http_error.status == 409:
+        raise WindowsAzureConflictError(
+            _ERROR_CONFLICT.format(str(http_error)))
+    elif http_error.status == 404:
+        raise WindowsAzureMissingResourceError(
+            _ERROR_NOT_FOUND.format(str(http_error)))
+    else:
+        if http_error.respbody is not None:
+            raise WindowsAzureError(
+                _ERROR_UNKNOWN.format(str(http_error)) + '\n' + \
+                    http_error.respbody.decode('utf-8-sig'))
+        else:
+            raise WindowsAzureError(_ERROR_UNKNOWN.format(str(http_error)))
+
+
+def _validate_type_bytes(param_name, param):
+    if not isinstance(param, bytes):
+        raise TypeError(_ERROR_VALUE_SHOULD_BE_BYTES.format(param_name))
+
+
+def _validate_not_none(param_name, param):
+    if param is None:
+        raise TypeError(_ERROR_VALUE_NONE.format(param_name))
+
+
+def _dont_fail_on_exist(error):
+    ''' don't throw exception if the resource exists.
+    This is called by create_* APIs with fail_on_exist=False'''
+    if isinstance(error, WindowsAzureConflictError):
+        return False
+    else:
+        raise error
+
+
+def _dont_fail_not_exist(error):
+    ''' don't throw exception if the resource doesn't exist.
+    This is called by create_* APIs with fail_on_exist=False'''
+    if isinstance(error, WindowsAzureMissingResourceError):
+        return False
+    else:
+        raise error
+
+
