+++ azure-servicebus/azure/servicebus/_common_serialization.py
@@ -14,8 +14,6 @@
 #--------------------------------------------------------------------------
 import ast
 import base64
-import hashlib
-import hmac
 import sys
 import types
 import warnings
@@ -24,12 +22,10 @@ if sys.version_info < (3,):
     from cStringIO import StringIO
     from urllib2 import quote as url_quote
     from urllib2 import unquote as url_unquote
-    _strtype = basestring
 else:
     from io import StringIO
     from urllib.parse import quote as url_quote
     from urllib.parse import unquote as url_unquote
-    _strtype = str
 
 from datetime import datetime
 from xml.sax.saxutils import escape as xml_escape
@@ -39,124 +35,28 @@ try:
 except ImportError:
     from xml.etree import ElementTree as ETree
 
-from azure.common import (
-    WindowsAzureError,
-    WindowsAzureConflictError,
-    WindowsAzureMissingResourceError,
-    WindowsAzureBatchOperationError,
-    WindowsAzureAsyncOperationError,
+from ._common_conversion import (
+    _str,
+)
+from ._common_error import (
+    _ERROR_VALUE_SHOULD_BE_BYTES,
+    _WARNING_VALUE_SHOULD_BE_BYTES,
+)
+from ._common_models import (
+    Feed,
+    HeaderDict,
+    WindowsAzureData,
+    _Base64String,
+    _dict_of,
+    _list_of,
+    _scalar_list_of,
+    _unicode_type,
+    _xml_attribute,
 )
 
-#--------------------------------------------------------------------------
-# constants
-
-__author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
-__version__ = '0.11.1'
-
-# Live ServiceClient URLs
-BLOB_SERVICE_HOST_BASE = '.blob.core.windows.net'
-QUEUE_SERVICE_HOST_BASE = '.queue.core.windows.net'
-TABLE_SERVICE_HOST_BASE = '.table.core.windows.net'
-SERVICE_BUS_HOST_BASE = '.servicebus.windows.net'
-MANAGEMENT_HOST = 'management.core.windows.net'
-
-# Development ServiceClient URLs
-DEV_BLOB_HOST = '127.0.0.1:10000'
-DEV_QUEUE_HOST = '127.0.0.1:10001'
-DEV_TABLE_HOST = '127.0.0.1:10002'
-
-# Default credentials for Development Storage Service
-DEV_ACCOUNT_NAME = 'devstoreaccount1'
-DEV_ACCOUNT_KEY = 'Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw=='
-
-# Default timeout for http requests (in secs)
-DEFAULT_HTTP_TIMEOUT = 65
-
 # All of our error messages
-_ERROR_CANNOT_FIND_PARTITION_KEY = 'Cannot find partition key in request.'
-_ERROR_CANNOT_FIND_ROW_KEY = 'Cannot find row key in request.'
-_ERROR_INCORRECT_TABLE_IN_BATCH = \
-    'Table should be the same in a batch operations'
-_ERROR_INCORRECT_PARTITION_KEY_IN_BATCH = \
-    'Partition Key should be the same in a batch operations'
-_ERROR_DUPLICATE_ROW_KEY_IN_BATCH = \
-    'Row Keys should not be the same in a batch operations'
-_ERROR_BATCH_COMMIT_FAIL = 'Batch Commit Fail'
-_ERROR_MESSAGE_NOT_PEEK_LOCKED_ON_DELETE = \
-    'Message is not peek locked and cannot be deleted.'
-_ERROR_MESSAGE_NOT_PEEK_LOCKED_ON_UNLOCK = \
-    'Message is not peek locked and cannot be unlocked.'
-_ERROR_EVENT_HUB_NOT_FOUND = 'Event hub was not found'
-_ERROR_QUEUE_NOT_FOUND = 'Queue was not found'
-_ERROR_TOPIC_NOT_FOUND = 'Topic was not found'
-_ERROR_CONFLICT = 'Conflict ({0})'
-_ERROR_NOT_FOUND = 'Not found ({0})'
-_ERROR_UNKNOWN = 'Unknown error ({0})'
-_ERROR_SERVICEBUS_MISSING_INFO = \
-    'You need to provide servicebus namespace, access key and Issuer'
-_ERROR_STORAGE_MISSING_INFO = \
-    'You need to provide an account name'
-_ERROR_ACCESS_POLICY = \
-    'share_access_policy must be either SignedIdentifier or AccessPolicy ' + \
-    'instance'
-_WARNING_VALUE_SHOULD_BE_BYTES = \
-    'Warning: {0} must be bytes data type. It will be converted ' + \
-    'automatically, with utf-8 text encoding.'
-_ERROR_VALUE_SHOULD_BE_BYTES = '{0} should be of type bytes.'
-_ERROR_VALUE_NONE = '{0} should not be None.'
-_ERROR_VALUE_NEGATIVE = '{0} should not be negative.'
-_ERROR_CANNOT_SERIALIZE_VALUE_TO_ENTITY = \
-    'Cannot serialize the specified value ({0}) to an entity.  Please use ' + \
-    'an EntityProperty (which can specify custom types), int, str, bool, ' + \
-    'or datetime.'
-_ERROR_PAGE_BLOB_SIZE_ALIGNMENT = \
-    'Invalid page blob size: {0}. ' + \
-    'The size must be aligned to a 512-byte boundary.'
-_ERROR_ASYNC_OP_FAILURE = 'Asynchronous operation did not succeed.'
-_ERROR_ASYNC_OP_TIMEOUT = 'Timed out waiting for async operation to complete.'
-
-_USER_AGENT_STRING = 'pyazure/' + __version__
 
 METADATA_NS = 'http://schemas.microsoft.com/ado/2007/08/dataservices/metadata'
-
-
-class WindowsAzureData(object):
-
-    ''' This is the base of data class.
-    It is only used to check whether it is instance or not. '''
-    pass
-
-
-class Feed(object):
-    pass
-
-
-class _Base64String(str):
-    pass
-
-
-class HeaderDict(dict):
-
-    def __getitem__(self, index):
-        return super(HeaderDict, self).__getitem__(index.lower())
-
-
-def _encode_base64(data):
-    if isinstance(data, _unicode_type):
-        data = data.encode('utf-8')
-    encoded = base64.b64encode(data)
-    return encoded.decode('utf-8')
-
-
-def _decode_base64_to_bytes(data):
-    if isinstance(data, _unicode_type):
-        data = data.encode('utf-8')
-    return base64.b64decode(data)
-
-
-def _decode_base64_to_text(data):
-    decoded_bytes = _decode_base64_to_bytes(data)
-    return decoded_bytes.decode('utf-8')
 
 
 _etree_entity_feed_namespaces = {
@@ -197,60 +97,13 @@ def _get_readable_id(id_name, id_prefix_to_skip):
     return id_name
 
 
-def _create_entry(entry_body):
-    ''' Adds common part of entry to a given entry body and return the whole
-    xml. '''
-    updated_str = datetime.utcnow().isoformat()
-    if datetime.utcnow().utcoffset() is None:
-        updated_str += '+00:00'
-
-    entry_start = '''<?xml version="1.0" encoding="utf-8" standalone="yes"?>
-<entry xmlns:d="http://schemas.microsoft.com/ado/2007/08/dataservices" xmlns:m="http://schemas.microsoft.com/ado/2007/08/dataservices/metadata" xmlns="http://www.w3.org/2005/Atom" >
-<title /><updated>{updated}</updated><author><name /></author><id />
-<content type="application/xml">
-    {body}</content></entry>'''
-    return entry_start.format(updated=updated_str, body=entry_body)
-
-
 def _to_datetime(strtime):
     return datetime.strptime(strtime, "%Y-%m-%dT%H:%M:%S.%f")
 
 _KNOWN_SERIALIZATION_XFORMS = {
-    'include_apis': 'IncludeAPIs',
-    'message_id': 'MessageId',
-    'content_md5': 'Content-MD5',
     'last_modified': 'Last-Modified',
     'cache_control': 'Cache-Control',
-    'account_admin_live_email_id': 'AccountAdminLiveEmailId',
-    'service_admin_live_email_id': 'ServiceAdminLiveEmailId',
-    'subscription_id': 'SubscriptionID',
-    'fqdn': 'FQDN',
-    'private_id': 'PrivateID',
-    'os_virtual_hard_disk': 'OSVirtualHardDisk',
-    'logical_disk_size_in_gb': 'LogicalDiskSizeInGB',
-    'logical_size_in_gb': 'LogicalSizeInGB',
-    'os': 'OS',
-    'persistent_vm_downtime_info': 'PersistentVMDowntimeInfo',
-    'copy_id': 'CopyId',
-    'os_state': 'OSState',
-    'vm_image': 'VMImage',
-    'vm_images': 'VMImages',
-    'os_disk_configuration': 'OSDiskConfiguration',
-    'public_ips': 'PublicIPs', 
-    'public_ip': 'PublicIP', 
-    'supported_os': 'SupportedOS',
-    'reserved_ip': 'ReservedIP',
-    'reserved_ips': 'ReservedIPs',
-    'aad_tenant_id': 'AADTenantID',
-    'start_ip_address': 'StartIPAddress',
-    'end_ip_address': 'EndIPAddress',
-    'operation_id': 'OperationId',
-    'operation_object_id': 'OperationObjectId',
-    'client_ip': 'ClientIP',
-    'status_id': 'ID',
-    'virtual_ips': 'VirtualIPs',
-    'virtual_ip': 'VirtualIP'
-    }
+}
 
 
 def _get_serialization_name(element_name):
@@ -268,45 +121,6 @@ def _get_serialization_name(element_name):
             element_name = element_name.replace('_', '-_')
 
     return ''.join(name.capitalize() for name in element_name.split('_'))
-
-if sys.version_info < (3,):
-    _unicode_type = unicode
-
-    def _str(value):
-        if isinstance(value, unicode):
-            return value.encode('utf-8')
-
-        return str(value)
-else:
-    _str = str
-    _unicode_type = str
-
-
-def _str_or_none(value):
-    if value is None:
-        return None
-
-    return _str(value)
-
-
-def _int_or_none(value):
-    if value is None:
-        return None
-
-    return str(int(value))
-
-
-def _bool_or_none(value):
-    if value is None:
-        return None
-
-    if isinstance(value, bool):
-        if value:
-            return 'true'
-        else:
-            return 'false'
-
-    return str(value)
 
 
 def _convert_class_to_xml(source, xml_prefix=True):
@@ -343,16 +157,6 @@ def _set_continuation_from_response_headers(feeds, response):
             x_ms_continuation[name[len('x-ms-continuation') + 1:]] = value
     if x_ms_continuation:
         setattr(feeds, 'x_ms_continuation', x_ms_continuation)
-
-
-def _validate_type_bytes(param_name, param):
-    if not isinstance(param, bytes):
-        raise TypeError(_ERROR_VALUE_SHOULD_BE_BYTES.format(param_name))
-
-
-def _validate_not_none(param_name, param):
-    if param is None:
-        raise TypeError(_ERROR_VALUE_NONE.format(param_name))
 
 
 def _get_request_body_bytes_only(param_name, param_value):
@@ -397,195 +201,6 @@ def _get_request_body(request_body):
         return request_body.encode('utf-8')
 
     return request_body
-
-
-class _dict_of(dict):
-
-    """a dict which carries with it the xml element names for key,val.
-    Used for deserializaion and construction of the lists"""
-
-    def __init__(self, pair_xml_element_name, key_xml_element_name,
-                 value_xml_element_name):
-        self.pair_xml_element_name = pair_xml_element_name
-        self.key_xml_element_name = key_xml_element_name
-        self.value_xml_element_name = value_xml_element_name
-        super(_dict_of, self).__init__()
-
-
-class _list_of(list):
-
-    """a list which carries with it the type that's expected to go in it.
-    Used for deserializaion and construction of the lists"""
-
-    def __init__(self, list_type, xml_element_name=None):
-        self.list_type = list_type
-        if xml_element_name is None:
-            self.xml_element_name = list_type.__name__
-        else:
-            self.xml_element_name = xml_element_name
-        super(_list_of, self).__init__()
-
-
-class _scalar_list_of(list):
-
-    """a list of scalar types which carries with it the type that's
-    expected to go in it along with its xml element name.
-    Used for deserializaion and construction of the lists"""
-
-    def __init__(self, list_type, xml_element_name):
-        self.list_type = list_type
-        self.xml_element_name = xml_element_name
-        super(_scalar_list_of, self).__init__()
-        
-class _xml_attribute:
-    
-    """a accessor to XML attributes
-    expected to go in it along with its xml element name.
-    Used for deserialization and construction"""
-    
-    def __init__(self, xml_element_name):
-        self.xml_element_name = xml_element_name
-
-
-def _update_request_uri_query_local_storage(request, use_local_storage):
-    ''' create correct uri and query for the request '''
-    uri, query = _update_request_uri_query(request)
-    if use_local_storage:
-        return '/' + DEV_ACCOUNT_NAME + uri, query
-    return uri, query
-
-
-def _update_request_uri_query(request):
-    '''pulls the query string out of the URI and moves it into
-    the query portion of the request object.  If there are already
-    query parameters on the request the parameters in the URI will
-    appear after the existing parameters'''
-
-    if '?' in request.path:
-        request.path, _, query_string = request.path.partition('?')
-        if query_string:
-            query_params = query_string.split('&')
-            for query in query_params:
-                if '=' in query:
-                    name, _, value = query.partition('=')
-                    request.query.append((name, value))
-
-    request.path = url_quote(request.path, '/()$=\',')
-
-    # add encoded queries to request.path.
-    if request.query:
-        request.path += '?'
-        for name, value in request.query:
-            if value is not None:
-                request.path += name + '=' + url_quote(value, '/()$=\',') + '&'
-        request.path = request.path[:-1]
-
-    return request.path, request.query
-
-
-def _dont_fail_on_exist(error):
-    ''' don't throw exception if the resource exists.
-    This is called by create_* APIs with fail_on_exist=False'''
-    if isinstance(error, WindowsAzureConflictError):
-        return False
-    else:
-        raise error
-
-
-def _dont_fail_not_exist(error):
-    ''' don't throw exception if the resource doesn't exist.
-    This is called by create_* APIs with fail_on_exist=False'''
-    if isinstance(error, WindowsAzureMissingResourceError):
-        return False
-    else:
-        raise error
-
-
-def _general_error_handler(http_error):
-    ''' Simple error handler for azure.'''
-    if http_error.status == 409:
-        raise WindowsAzureConflictError(
-            _ERROR_CONFLICT.format(str(http_error)))
-    elif http_error.status == 404:
-        raise WindowsAzureMissingResourceError(
-            _ERROR_NOT_FOUND.format(str(http_error)))
-    else:
-        if http_error.respbody is not None:
-            raise WindowsAzureError(
-                _ERROR_UNKNOWN.format(str(http_error)) + '\n' + \
-                    http_error.respbody.decode('utf-8-sig'))
-        else:
-            raise WindowsAzureError(_ERROR_UNKNOWN.format(str(http_error)))
-
-
-def _parse_response_for_dict(response):
-    ''' Extracts name-values from response header. Filter out the standard
-    http headers.'''
-
-    if response is None:
-        return None
-    http_headers = ['server', 'date', 'location', 'host',
-                    'via', 'proxy-connection', 'connection']
-    return_dict = HeaderDict()
-    if response.headers:
-        for name, value in response.headers:
-            if not name.lower() in http_headers:
-                return_dict[name] = value
-
-    return return_dict
-
-
-def _parse_response_for_dict_prefix(response, prefixes):
-    ''' Extracts name-values for names starting with prefix from response
-    header. Filter out the standard http headers.'''
-
-    if response is None:
-        return None
-    return_dict = {}
-    orig_dict = _parse_response_for_dict(response)
-    if orig_dict:
-        for name, value in orig_dict.items():
-            for prefix_value in prefixes:
-                if name.lower().startswith(prefix_value.lower()):
-                    return_dict[name] = value
-                    break
-        return return_dict
-    else:
-        return None
-
-
-def _parse_response_for_dict_filter(response, filter):
-    ''' Extracts name-values for names in filter from response header. Filter
-    out the standard http headers.'''
-    if response is None:
-        return None
-    return_dict = {}
-    orig_dict = _parse_response_for_dict(response)
-    if orig_dict:
-        for name, value in orig_dict.items():
-            if name.lower() in filter:
-                return_dict[name] = value
-        return return_dict
-    else:
-        return None
-
-
-def _sign_string(key, string_to_sign, key_is_base64=True):
-    if key_is_base64:
-        key = _decode_base64_to_bytes(key)
-    else:
-        if isinstance(key, _unicode_type):
-            key = key.encode('utf-8')
-    if isinstance(string_to_sign, _unicode_type):
-        string_to_sign = string_to_sign.encode('utf-8')
-    signed_hmac_sha256 = hmac.HMAC(key, string_to_sign, hashlib.sha256)
-    digest = signed_hmac_sha256.digest()
-    encoded_digest = _encode_base64(digest)
-    return encoded_digest
-
-
-def _lower(text):
-    return text.lower()
 
 
 class _ETreeXmlToObject(object):
